# src/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Kay\u0131t i\u015flemi",
                "description": "Kay\u0131t i\u015flemi",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "email adresi",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User e-mail address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Kullan\u0131c\u0131 \u015fifresi",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Kay\u0131t Ba\u015far\u0131l\u0131",
                        "schema": {
                            "properties": {
                                "errorCode": {
                                    "type": "int"
                                },
                                "errorMessage": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login i\u015flemi",
                "description": "Login i\u015flemi",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User e-mail address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Kay\u0131t Ba\u015far\u0131l\u0131",
                        "schema": {
                            "properties": {
                                "errorCode": {
                                    "type": "int"
                                },
                                "errorMessage": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout i\u015flemi",
                "description": "Login i\u015flemi",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Kay\u0131t Ba\u015far\u0131l\u0131",
                        "schema": {
                            "properties": {
                                "errorCode": {
                                    "type": "int"
                                },
                                "errorMessage": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profil bilgisi",
                "description": "Profil bilgisi",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "User token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "profile data",
                        "schema": {
                            "properties": {
                                "user_id": {
                                    "type": "integer"
                                },
                                "name_surname": {
                                    "type": "string"
                                },
                                "age": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}